requirements.txt
gnureadline==6.3.3
ipython==2.4.1
nose==1.3.4
pycrypto==2.6.1
